Move Attributes (nodes, edges) to seperate header.
